Patch by F. E.: Fix before-use ref invalidation in PdfPage::DeleteAnnotation(const PdfReference &)
This fix is from a post to the podofo-users mailing list (see bottom
of commit message for metadata) in its attachment. It moves the erase()
on the annotations array for the annotation to be deleted the end of the
method to keep its ref (given to the method) valid to delete it from the
cache (if it's there) and from the document (reported broken before).   

Post-Author: F. E. <exler7287@gmail.com>
Post-Subject: Re: [Podofo-users] PdfPage::DeleteAnnotation does not delete the annotation object as it should (full mail)
Post-Archive-Date: 2019-04-15 07:54:54 +0000
Post-Message-ID: CAFKhaC%2BbOGevxwCkx2_5M5X8ug0kX2bxKF71ciz8NTJM8k%3D%3Dfg@mail.gmail.com
Post-Archive-URL: https://sourceforge.net/p/podofo/mailman/message/36640320/


git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@1980 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5